Update class.osticket.php
I don't know if this is the osTicket way. I'm hoping someone will correct me. My problem was that when a plugin loaded it it created all the links relative to the plugin url. Instead of the /scp/dashboard.php, I'd get /scp/myplugin-directory/dashboard.php. That does not work as you know.

I created a new variable, setter, and getter function in class.osticket.php.

$plugin_instance is set to false because the default it not to be a plugin.
setPluginInstance($plugin_path) sets the path of the plugin.
isPlugin returns the path value.

isPlugin can be used to hold the path (or anything really) that can be placed in a template.  setPluginInstance is called from a controller method and $plugin_instance is set with the path desired.

I used this in staff/templates/navigation.tmpl.php like so:

echo sprintf('<li class="%s %s"><a href="'.  $ost->isPlugin()    .'%s">%s</a>',

and

 '<li><a class="%s" href="'.  $ost->isPlugin()    .'%s" title="%s" id="%s">%s</a></li>',